Revert "Merge branch '2.3.0' into csharp/client_refactoring"
This reverts commit d77c453e7a7b78ddc11759959f205c90fbff2801, reversing
changes made to da8651a74ba7f9c62dedfb95197f79b93b41c036.

+++ samples/client/petstore/tizen/src/User.h
@@ -1,7 +1,23 @@
 /*
+ * Copyright (c) 2016 Samsung Electronics Co., Ltd All Rights Reserved
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
  * User.h
  *
- * A User who is purchasing from the pet store
+ * 
  */
 
 #ifndef _User_H_
@@ -15,7 +31,7 @@ namespace Tizen {
 namespace ArtikCloud {
 
 
-/*! \brief A User who is purchasing from the pet store
+/*! \brief 
  *
  */
 
@@ -25,7 +41,7 @@ public:
 	 */
 	User();
 	User(char* str);
-
+	
 	/*! \brief Destructor.
 	 */
 	virtual ~User();
@@ -40,11 +56,11 @@ public:
 
 	/*! \brief Get 
 	 */
-	long long getId();
+	long getId();
 
 	/*! \brief Set 
 	 */
-	void setId(long long  id);
+	void setId(long  id);
 	/*! \brief Get 
 	 */
 	std::string getUsername();
@@ -96,7 +112,7 @@ public:
 	void setUserStatus(int  userStatus);
 
 private:
-	long long id;
+	long id;
 	std::string username;
 	std::string firstName;
 	std::string lastName;
@@ -106,7 +122,7 @@ private:
 	int userStatus;
 	void __init();
 	void __cleanup();
-
+	
 };
 }
 }
